requirements.txt
asgiref==3.9.2
attrs==25.4.0
certifi==2026.1.4
Django==5.2.7
django-debug-toolbar==6.2.0
django-filter==25.2
djangorestframework==3.16.1
docutils==0.22.4
drf-spectacular==0.29.0
flake8==7.3.0
flake8-docstrings==1.7.0
gunicorn==25.0.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
packaging==26.0
pillow==12.1.0
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sentry-sdk==2.51.0
snowballstemmer==3.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
